Docs (link): Update links for individual builds. Closes #7533.
Previously, the custom builds documentation instructed to clone the
ckeditor5-build-classic (and variants) repositories. Since the
inclusion of these repository under the main monorepo, the individual
repositories have been marked "archived" and read-only.

This commit modifies the instructions to instruct cloning the main
monorepo and working directly out of the subdirectory.