update external ui doc
When I tried to render a `ContextualBalloon` for the `Mention` plugin, it was not rendering because the custom editor couldn't find the view in the document to attach to. 

I'm not sure whether this is the correct way, the purpose of this PR is to confirm whether this was a bug, or is there a better way so that `ContextualBalloon` itself will call the `attachToDom` function.