Fix (table): Support cells without children
Previously, it was possible for table cells to be entirely empty.
However, a recent change introduced in 93dbbf99 changed the evaluation order for table cell children.

This led to calling hasAnyAttribute on a null reference, which is not supported.

Move isSingleParagraph check prior to hasAnyAttribute. isSingleParagraph is correctly calculated because it relies on the childCount for the table cell.

Fixes #8941
Fixes #8917