Update angular.md
Based on comments in https://github.com/ckeditor/ckeditor5-angular/issues/211#issuecomment-644862143 it sounds like the watchdog is already added to the Angular component. It's possible this entire section needs to be removed, however, I wanted to add a note that would prevent future users from trying to add this ContextWatchdog to their own components. FWIW, I also receive an error stating the ContextWatchdog method is not a constructor, which led me to discovering the comment. 

Thanks in advance for you time and everything you all do with CKEditor! Love the editor.