Fix crash in renderer
This loop was doing unsafe removals of DOM nodes while iterating over indices that point into a live `NodeList`, and was causing a crash in certain cases.

The error is:

```
TypeError: undefined is not iterable (cannot read property Symbol(Symbol.iterator))
    at Function.from (<anonymous>)
    at Renderer._updateAttrs (webpack:///./packages/ckeditor5-engine/src/view/renderer.js?:19:11593)
    at Renderer.render (webpack:///./packages/ckeditor5-engine/src/view/renderer.js?:19:2390)
```

and it can happen when this loop ends up setting up an incorrect mapping from a view element to a text node, and then `_updateAttrs()` is called on that text node.

The factors that I've identified that will cause this to reproduce are:
* Enough nodes to make `_diffNodeLists()` use the fastDiff implementation
* A render involving at least two changes (e.g. removals) separated other nodes