JSDoc: Add DOMRect to getOptimalPosition and Rect
The constructor for `Rect` shows this example:

`document.body.getClientRects().item( 0 )`

The return type in every browser but IE is DOMRect so I'm adding it to
the possible values of the `Rect` constructor. I'm also adding `Window`
cause `Rect` can handle it.

For the `target` property of `Options` interface, I'm adding the same
types cause that property is used to create a new `Rect`.