change default text inside of an inserted resource link to resource title
This adds a second `title` argument to the command for creating a
resource link, and in the case where we're creating a link from scratch
(rather than adding the link properties to existing text) we use that as
the text inside that link. Formerly we were using the UUID there, which
isn't the most friendly thing to do.

pr #8