Moved color conversion handling from private _hexColor handler to color public prop change.
There was an edge case with former solution:

At one point color property (format conversion) was driven by _hexColor value **change**.

This was wrong because _hexColor is normalized, so if you set it to same color in one format

and then in another - there's no change in normalized color.

So _hexColor change event was not triggered, thus no format was enforced.