Allow to turn a particular root to read-only mode (#13726)
Feature (editor-multi-root): Introduced `MultiRootEditor#disableRoot()` and `enableRoot()` that allows to switch only a particular root to a read-only state.

Feature (engine): Introduced `Model#canEditAt( selectable )` method that checks whether given `selectable` is at an editable place in the model. It should be used instead of `Editor#isReadOnly` in multi-root integrations as it acknowledges the roots read-only state.