Optimize body wrapper lookup
On pages with lots of content and a large number of editors being created at once, the `querySelector` in `attachToDom()` shows up as a non-trivial in performance profiling. The `ck-body-wrapper` element is always added as a direct child of the `body`, so we can speed up the lookup a lot by just iterating over the children of the `body` rather than running `querySelector()`.